OD-10844 fix id type
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@49208 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ editor-api/src/main/groovy/ish/oncourse/willow/editor/service/impl/PageApiServiceImpl.groovy
@@ -27,9 +27,8 @@ class PageApiServiceImpl implements PageApi {
     Page addPage() {
         WebNode node = WebNodeFunctions.createNewNode(requestService.request, cayenneService.newContext())
         new Page().with {page ->
-            page.id = node.id
+            page.id = node.id.doubleValue()
             page.title = node.name
-
             page
         }
     }
@@ -56,7 +55,7 @@ class PageApiServiceImpl implements PageApi {
 
         nodes.collect {node ->
             new Page().with {page ->
-                page.id = node.id
+                page.id = node.id.doubleValue()
                 page.title = node.name
                 page.visible = node.published
                 page.urls = node.webUrlAliases.collect {urlAlias ->
